requirements.txt
Django==1.6.5
Pillow==2.4.0
django-filebrowser-no-grappelli==3.1.1
django-fluent-pages==0.8.6
django-mptt==0.6.1
django-polymorphic==0.5.5
django-polymorphic-tree==0.8.11
django-tag-parser==1.1.0
django-tinymce==1.5.2
psycopg2==2.5.3
django-disqus==0.4.1
South==1.0
django-simple-captcha==0.4.2